# swagger_server/models/bank_transfer_payment.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.payment import Payment  # noqa: F401,E501
from swagger_server import util


class BankTransferPayment(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, name: str=None, payment: Payment=None):  # noqa: E501
        """BankTransferPayment - a model defined in Swagger

        :param name: The name of this BankTransferPayment.  # noqa: E501
        :type name: str
        :param payment: The payment of this BankTransferPayment.  # noqa: E501
        :type payment: Payment
        """
        self.swagger_types = {
            'name': str,
            'payment': Payment
        }

        self.attribute_map = {
            'name': 'name',
            'payment': 'payment'
        }
        self._name = name
        self._payment = payment

    @classmethod
    def from_dict(cls, dikt) -> 'BankTransferPayment':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BankTransferPayment of this BankTransferPayment.  # noqa: E501
        :rtype: BankTransferPayment
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this BankTransferPayment.


        :return: The name of this BankTransferPayment.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this BankTransferPayment.


        :param name: The name of this BankTransferPayment.
        :type name: str
        """

        self._name = name

    @property
    def payment(self) -> Payment:
        """Gets the payment of this BankTransferPayment.


        :return: The payment of this BankTransferPayment.
        :rtype: Payment
        """
        return self._payment

    @payment.setter
    def payment(self, payment: Payment):
        """Sets the payment of this BankTransferPayment.


        :param payment: The payment of this BankTransferPayment.
        :type payment: Payment
        """

        self._payment = payment